% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utils.R
\name{add_duplicated_flag}
\alias{add_duplicated_flag}
\title{Add Duplicated Observations Flag}
\usage{
add_duplicated_flag(data, dupvars)
}
\arguments{
\item{data}{Data frame to check.}

\item{dupvars}{Character vector providing the names of variables for which duplicated observations should be checked.}
}
\value{
The data frame \code{data}, with an additional column \code{"dupobs"} that is a logical indicator of a duplicated row.
}
\description{
Adds a column indicating duplicated rows
}
\details{
This is a wrapper around \code{\link{duplicated}}.
}
