% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utils.R
\name{add_Ht_standata}
\alias{add_Ht_standata}
\alias{add_Hx_standata}
\title{Add time spline to a standata object}
\usage{
add_Ht_standata(standata, Ht)

add_Hx_standata(standata, Hx)
}
\arguments{
\item{standata}{the standata object to be updated}

\item{Ht, Hx}{the matrix of time or exposure splines to be added}
}
\description{
Updates a standata object with a matrix of time or exposure splines
}
\details{
\code{add_Ht_standata} updates the \code{Ht}, \code{timedf}, and \code{Ht_attr} elements of \code{standata}. \code{add_Hx_standata} updates the \code{Hx}, \code{xdf}, and \code{Hx_attr} elements of \code{standata}.
}
\seealso{
\code{\link{create_standata_exposure}}, \code{\link{create_standata_outcome}}, \code{\link{expsim_update_parameter}}, \code{\link{outsim_update_parameter}}, \code{\link{create_spline_range}}
}
