% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stan_objects.R
\name{add_priors}
\alias{add_priors}
\alias{add_priors_exposure_model}
\alias{add_priors_outcome_model}
\title{Add Hyperparameter Values to Standata Object}
\usage{
add_priors(standata, ...)

add_priors_exposure_model(
  standata,
  etaG = 0,
  sigmaG = c(0, 1),
  reK = 0,
  sigmaK = c(0, 1),
  reI = 0,
  sigmaI = c(0, 1),
  sigmaW = c(0, 1),
  theta = 0,
  sigmaTheta = c(0, 1),
  ...
)

add_priors_outcome_model(
  standata,
  beta = 0,
  sigmaBeta = c(0, 1),
  gamma = 0,
  sigmaGamma = c(0, 1),
  delta = 0,
  sigmaDelta = c(0, 1),
  sigmaI = c(0, 1),
  beta_nu = 1,
  ...
)
}
\arguments{
\item{standata}{Either a \code{standata_exposure} or \code{standata_outcome} object. Based on its class, \code{add_priors} calls \code{add_priors_exposure_model} or \code{add_priors_exposure_model}.}

\item{...}{Arguments passed to \code{add_priors_exposure_model} or \code{add_priors_outcome_model}. Unlisted arguments are ignorred.}

\item{etaG}{Prior mean and standard deviation for the group means parameter `etaG`.}

\item{sigmaG}{Prior mean and standard deviation for the group standard deviation parameter `sigmaG`.}

\item{reK}{Prior mean for cluster-level random effects.}

\item{sigmaK}{Prior mean and standard deviation for the standard deviation of the cluster random effects, `sigmaK`.}

\item{reI}{Prior mean for household-level random effects.}

\item{sigmaI}{Prior mean and standard deviation for the standard deviation of the subject-level random effect.}

\item{sigmaW}{Prior mean and standard deviation for the residual standard deviation parameter `sigmaW`.}

\item{theta}{Prior mean and standard deviation for the time coefficients `thetaG`.}

\item{sigmaTheta}{Prior mean and standard deviation for the standard deviation parameter for time trends, `sigmaTheta`.}

\item{beta}{Prior mean for the exposure coefficient(s).}

\item{sigmaBeta}{Prior mean and standard deviation for the standard deviation of exposure coefficient parameter.}

\item{gamma}{Prior mean for the covariate coefficient(s).}

\item{sigmaGamma}{Prior mean and standard deviation for the standard deviation of covariate coefficient parameter.}

\item{delta}{Prior mean fro the time coefficient(s).}

\item{sigmaDelta}{Prior mean and standard deviation for the standard deviation of time coefficient parameter.}

\item{beta_nu}{Prior value for the LKJ prior on the correlation between exposure coefficients from different studies.}
}
\description{
Sets prior distribution hyperparameters using defaults or provided values
}
\details{
For \code{add_stan_priors_exposure_model}, the prior distributions are normal distributions parameterixed by (mean, sd).
}
\seealso{
\code{\link{create_standata_exposure}}
}
