% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utils.R
\name{add_spline_time}
\alias{add_spline_time}
\alias{add_spline_exposure}
\title{Add spline to a standata object}
\usage{
add_spline_time(standata, Mt)

add_spline_exposure(standata, Mx)
}
\arguments{
\item{standata}{the standata object to be updated}

\item{Mt, Mx}{the matrix of time or exposure splines to be added}
}
\description{
Updates a standata object with a matrix of time or exposure splines
}
\details{
\code{add_spline_time} updates the \code{Mt}, \code{timedf}, and \code{Mt_attributes} elements of \code{standata}. \code{add_spline_exposure} updates the \code{Mx}, \code{xdf}, and \code{Mx_attributes} elements of \code{standata}.
}
\seealso{
\code{\link{create_standata_exposure}}, \code{\link{create_standata_outcome}}, \code{\link{expsim_update_parameter}}, \code{\link{outsim_update_parameter}}, \code{\link{create_spline_range}}
}
