% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bercs-package.R
\docType{package}
\name{bercs-package}
\alias{bercs-package}
\alias{bercs}
\title{The 'bercs' package.}
\description{
Bayesian Exposure-Response Curves via STAN
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
