% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{casedataE}
\alias{casedataE}
\title{Example data giving an exponential relationship between conc and sbp}
\format{
A data frame with 200 rows and 6 columns:
\describe{
\item{conc}{concentration of exposure}
\item{sbp}{systolic blood pressure}
\item{unit_id}{the id corresponding to an individual from a study}
\item{study}{the study from which the observation was obtained}
}
}
\usage{
casedataE
}
\description{
Example data giving an exponential relationship between conc and sbp
}
\examples{
data(casedataE)
plot(x=casedataE$conc, y=casedataE$sbp)
}
\keyword{datasets}
