% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{casedataF}
\alias{casedataF}
\title{Example data giving a piecewise relationship between conc and sbp}
\format{
A data frame with 200 rows and 6 columns:
\describe{
  \item{conc}{concentration of exposure}
  \item{sbp}{systolic blood pressure}
  \item{unit_id}{the id corresponding to an individual from a study}
  \item{study}{the study from which the observation was obtained}
}
}
\usage{
casedataF
}
\description{
Example data giving a piecewise relationship between conc and sbp
}
\examples{
data(casedataF)
plot(x=casedataF$conc, y=casedataF$sbp)
}
\keyword{datasets}
