% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_outcome_model.R
\name{compute_ERC}
\alias{compute_ERC}
\alias{plot_ERC}
\alias{center_ERC}
\alias{compute_OR}
\title{Extract fitted Exposure-Response Curve (ERC)}
\usage{
compute_ERC(
  standata,
  stanfit,
  exprange = c(0, 100),
  expsequence = NULL,
  ref_exposure = NULL,
  ciband = 0.95,
  inclInterceptUncertainty = TRUE,
  inclIntercept = FALSE,
  intercept_prop = c("equal", "obs"),
  study = NULL,
  beta_post,
  bs_post,
  nS = standata$S,
  Mx = standata$Mx,
  Mx_attributes = standata$Mx_attributes,
  xdf = standata$xdf,
  ...
)

plot_ERC(
  obj,
  incS = NULL,
  expERC = TRUE,
  ylab = "Relative Risk",
  xlab = "Exposure",
  ribbon = FALSE,
  hline = NULL
)

center_ERC(obj, ref_exposure = min(obj$exposure))

compute_OR(
  standata,
  stanfit,
  exprange = c(0, 100),
  expsequence = NULL,
  ref_exposure = 0,
  ciband = 0.95,
  inclInterceptUncertainty = TRUE,
  inclIntercept = FALSE,
  intercept_prop = c("equal", "obs"),
  study = NULL,
  beta_post,
  bs_post,
  nS = standata$S,
  Mx = standata$Mx,
  Mx_attributes = standata$Mx_attributes,
  xdf = standata$xdf,
  ...
)
}
\arguments{
\item{standata}{\code{standata_outcome} object used to fit the model}

\item{stanfit}{fitted outcome model from \code{\link{sample_outcome_model}}. Posterior samples of 'beta' are extracted from this object.}

\item{exprange}{range of exposure values over which to compute the curve}

\item{expsequence}{sequence of exposure values at which the curve should be evaluated. For plotting, it is preferable to use \code{exprange}, but for specific known exposure values use \code{expsequence}.}

\item{ref_exposure}{Exposure concentration at which the spline should be shifted to have value 0 on log scale (value 1 on exponentiated scale). The largest exposure value less than or equal to \code{ref_exposure} is used as the new reference concentration.}

\item{ciband}{width of credible interval band}

\item{inclInterceptUncertainty}{Should intercept uncertainty be included uncertainty estimates? See details for more information.}

\item{inclIntercept}{Should the intercept term be included in the curve?}

\item{intercept_prop}{Proportions used in calculating the "average" intercept. Defaults to equal proportions for each study (\code{"equal"}) and can be set to be proportional to the number of observations in each study (\code{"obs"}). A vector of proportions can also be given.}

\item{study}{Return the curve for a specific study, or the}

\item{beta_post}{vector or matrix of posterior samples for 'beta' parameter. Only needed if \code{stanfit} not provided.}

\item{bs_post}{vector or matrix of posterior samples for 'bS' parameter. Only needed if \code{stanfit} not provided.}

\item{nS}{Number of studies. Only needed if \code{standata} not provided.}

\item{Mx}{Spline matrix for exposure. Only needed if \code{standata} not provided.}

\item{Mx_attributes}{Attributes for the spline matrix. Only needed if \code{standata} not provided.}

\item{xdf}{Degrees of freedom in exposure splines}

\item{...}{Passed to \code{\link{seq}} to control sequence of exposure values.}

\item{obj}{Data frame containing \code{exposure}, \code{mean}, \code{low}, and \code{high}. Typically generated from \code{\link{compute_ERC}}.}

\item{incS}{If model has a single curve but with different selections of intercept uncertainty, this indicates which choice of uncertainty to use. Defaults to the last column of the curve in \code{obj}, which is typically the averaged intercept uncertainty. If multiple curves are fit, this selects which curve(s) is plotted.}

\item{expERC}{Should the fitted curve be exponentiated (TRUE) or not (FALSE).}

\item{ylab}{String providing y-axis label.}

\item{xlab}{String providing x-axis label.}

\item{ribbon}{Should the uncertainty be represented as a filled ribbon (TRUE) or lines without fill (FALSE).}

\item{hline}{Optional argument that plots a horizontal line at given value.}
}
\description{
Computes the exposure-response curve from a fitted outcome model
}
\details{
This function creates a data frame containing the values of the exposure-response curve over a given range of values. The output is designed for easy plotting.
Currently, the fitted curve is plotted based upon the posterior means of the parameters, with the uncertainty bands based upon quantiles.

Uncertainty from the intercepts is included in the confidence bands by default, since this corresponds to a common interpretation of such intervals. This requires picking a single value for the intercept. For models fit to multiple studies, a separate set of uncertainty will be created for each study. Additionally, a set of results corresponding to "average" intercept is created. The \code{intercept_prop} argument controls the relative contribution of the intercepts from each model.

To calculate odds ratios for specific combinations of exposure values, use the \code{compute_OR} function, which will correctly calculate the credible intervals for the relative difference.
}
\examples{

#Load needed data sets
data(casedataA)
data(casedataB)

#Create standata
outcome_combo_data <- create_standata_outcome(datalist=list(casedataA, casedataB),
                                              xdf=4,
                                              xfnargs=list(Boundary.knots=c(5, 200)))

#Add priors
outcome_combo_data <- add_priors(outcome_combo_data,
sigmaI=c(0, 0.01))

#Sample from posterior distribution
outcome_combo_mod_fit <- sample_outcome_model(outcome_combo_data,
B=200,
cores=1)

#Compute and fit the exposure response curve
fitted_ERC <- compute_ERC(standata=outcome_combo_data,
stanfit=outcome_combo_mod_fit,
exprange=c(5,200))
plot_ERC(fitted_ERC) + ggplot2::scale_y_log10()

#Calculate odds ratio
compute_OR(standata=outcome_combo_data,
stanfit=outcome_combo_mod_fit,
expsequence = c(5, 10, 20, 50, 100, 200),
ref_exposure=10)

center_ERC(fitted_ERC, ref_exposure = min(fitted_ERC$exposure))

}
\seealso{
\code{\link{sample_outcome_model}}
}
