% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exposure_model.R
\name{compute_fitted_mean}
\alias{compute_fitted_mean}
\title{Compute estimated concentration}
\usage{
compute_fitted_mean(
  stanfit,
  standata,
  parvalues = NULL,
  include_time = TRUE,
  include_reI = TRUE,
  exp_transform = FALSE,
  add = FALSE,
  ...
)
}
\arguments{
\item{stanfit}{Fitted STAN model object containing posterior samples of parameters; from \code{\link{sample_exposure_model}}.}

\item{standata}{List containing model structure information corresponding to \code{stanfit}.
Typically a list of class \code{standata_exposure}. However, it may be any list (or data frame) containing
\code{cluster_of_obs} (if \code{etaK} was sampled in \code{stanfit}), \code{group_of_obs}
(if \code{etaK} was not sampled in \code{stanfit}),  \code{unit_of_obs} (if \code{include_reI=TRUE}),
 and \code{Mt} (if \code{include_time=TRUE}). This may be useful when calculating fitted means for records
  at different times from the observations used to fit the model.}

\item{parvalues}{List of parameter values to use for computing means. Not needed if \code{stanfit} is provided.}

\item{include_time}{Logical indicator of whether time should be included.}

\item{include_reI}{Logical indicator of whether the unit-level random effect should be included.}

\item{exp_transform}{Logical indicator of whether concentrations should be exponentiated.}

\item{add}{Logical indicator of whether the modeled means should be added to \code{standata} or just returned directly (the default).}

\item{...}{Additional arugments passed to \code{\link[rstan]{extract}}.}
}
\description{
Calculates the posterior mean exposure concentrations.
}
\examples{
#Create simulated data
exposure_standata <- create_standata_exposure(group=rep(1:3, each=24),
conc=rnorm(n=72,mean=rep(c(0, 2, 4), each=23)),
unit_id=rep(1:18, each=4),
time=rep(1:4, times=18))

#Add natural splines with 3 df
Mt <- create_spline_matrix(x=exposure_standata$time, df=3, fn="ns")

#Add priors and spline
exposure_standata <- add_spline_time(exposure_standata, Mt=Mt)
exposure_standata <- add_priors(exposure_standata, sigmaI=c(0, 0.5))

#Sample from posterior distribution using STAN
exposure_mod_fit <- sample_exposure_model(exposure_standata, B=200, chains=2)

#Compute long term means
compute_fitted_mean(stanfit=exposure_mod_fit, standata=exposure_standata)
}
\seealso{
\code{\link{plot_exposure_means_bytime}}, \code{\link{sample_exposure_model}}
}
