% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pooling.R
\name{compute_pooling_metrics}
\alias{compute_pooling_metrics}
\alias{prop_var_explained}
\alias{pooling_factor}
\title{Measures of Variance Explained and Pooling/Shrinkage}
\usage{
compute_pooling_metrics(theta, epsilon)

prop_var_explained(theta, epsilon)

pooling_factor(epsilon)
}
\arguments{
\item{theta}{Matrix of the 'outcome' values for the level. Rows represent samples from the 
posterior distribution and columns represent units of the level.}

\item{epsilon}{Matrix of the 'error' values for the level, in the same format as \code{theta}.}
}
\description{
Computes a form of R2 and Pooling Factor for a level of a hierarchical model
}
\details{
This computes generalized versions of R^2 and pooling factor for levels of a hierarchical model, as described by Gelman and Pardoe (2006).
}
\references{
Gelman, A and Pardoe, I. (2006). Bayesian Measures of Explained Variance and Pooling in Multilevel (Hierarchical) Models. \emph{Technometrics}, 48, 241-251.
}
\seealso{
\code{\link{exposure_pooling_factor}}
}
