% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utils.R
\name{create_spline}
\alias{create_spline}
\alias{create_spline_matrix}
\alias{predict_spline}
\title{Create spline matrix}
\usage{
create_spline(x, xrange = NULL, by = 1, df = 1, fn = "ns", ...)

create_spline_matrix(x, df = 1, fn = "ns", ...)

predict_spline(spline, newx, center = FALSE, scale = FALSE)
}
\arguments{
\item{x}{Variable values for which spline is to be created.}

\item{xrange}{Range (min and max) of values to use for creating spline. Only used if \code{x} not provided.}

\item{by}{Interval of generated sequence.}

\item{df}{Degrees of freedom for spline. If set to 1, then a linear function assumed; otherwise \code{df} is passed to \code{fn}.}

\item{fn}{Character string specifying the spline-generating function. Defaults to \code{\link[splines]{ns}}.}

\item{...}{additional arguments passed to \code{fn}.}

\item{spline}{Spline object from which to predict.}

\item{newx}{New values for which splines should be generated}

\item{center, scale}{Should spline values be centered and/or scaled? Default is \code{FALSE}. If a vector is provided, it is pased to \code{\link{scale}}.}
}
\value{
For \code{create_spline}, a list containing:
\describe{
\item{\code{x}}{the values provided in argument \code{x}}
\item{\code{spline}}{a matrix of spline values evaluated at \code{x}}
\item{\code{spline_mean}}{a vector of column-wise means of \code{spline}}
}

For \code{create_spline_matrix}, the matrix of spline values only.

For \code{predict_spline}, a matrix of spline values evaluated at \code{newx}.
}
\description{
Wrapper function for creating a spline matrix
}
\details{
Note that \code{predict_spline} relies on the existence of a \code{predict} generic for the class of spline.
}
\examples{
x_ns <- create_spline(x=1:10, df=3, fn="ns")
x2_ns <- predict(x_ns$spline, newx=seq(0.5, 9.5, by=1))
x2_ns
}
\seealso{
\code{\link{add_spline_time}}, \code{\link{sim_update_time_splines}}
}
