% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utils.R
\name{create_spline_matrix}
\alias{create_spline_matrix}
\alias{create_spline}
\alias{create_spline_range}
\alias{predict_spline}
\title{Create spline matrix}
\usage{
create_spline_matrix(x, df = 1, fn = "ns", ...)

create_spline(x, df = 1, fn = "ns", ...)

create_spline_range(df, xseq = NULL, xrange = NULL, by = 1, fn = "ns", ...)

predict_spline(spline, newx, center = FALSE, scale = FALSE)
}
\arguments{
\item{x}{Variable values for which spline is to be created.}

\item{df}{Degrees of freedom for spline. If set to 1, then a linear function assumed; otherwise \code{df} is passed to \code{fn}.}

\item{fn}{Character string specifying the spline-generating function. Defaults to \code{\link[splines]{ns}}.}

\item{...}{additional arguments passed to \code{fn}.}

\item{xseq}{Sequence of values to use for creating time spline.}

\item{xrange}{Range (min and max) of values to use for creating spline. Only used if \code{xseq} not provided.}

\item{by}{Interval of generated sequence.}

\item{spline}{Spline object from which to predict.}

\item{newx}{New values for which splines should be generated}

\item{center, scale}{Should spline values be centered and/or scaled? Default is \code{FALSE}. If a vector is provided, it is pased to \code{\link{scale}}.}
}
\description{
Wrapper function for creating a spline matrix
}
\seealso{
\code{\link{sim_update_time_splines}}
}
