% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exposure_model.R
\name{create_standata_exposure}
\alias{create_standata_exposure}
\title{Create List for Fitting Exposure Model via STAN}
\usage{
create_standata_exposure(data, Ht = NULL, log_transform = FALSE,
  return_addition = FALSE)
}
\arguments{
\item{data}{Data frame containing data in a long formet}

\item{Ht}{Matrix of time splines to include in the model. Defaults to NULL.}

\item{log_transform}{Should the concentration value be log-transformed?}

\item{return_addition}{See 'Value'.}
}
\value{
A list of class \code{standata_exposure} that contains the following:
\itemize{
\item {\code{G} -- Number of groups}
\item {\code{K} -- Number of cluster *nested within groups*.}
\item{\code{H} Number of households}
\item{\code{N} -- Number of observations}
\item{\code{group_of_cluster} -- Integer providing group number of each cluster}
\item{\code{cluster_of_obs} -- Integer providing cluster number of each observation}
\item{\code{group_of_obs} -- Integer providing group number of each observation}
\item{\code{hh_of_obs} -- Integer providing household number of each observation}
\item{\code{w} -- Concentration observations}
\item{\code{times} -- The date/time of each observation. This is not used directly in model fitting, but contained in the object to facilitate plotting and other summaries.}
}
If \code{return_addition=TRUE}, then a two-element list is returned. The first element
is the \code{standata_exposure} object described above. The second element is a modified version of \code{data}, with the variables \code{group_of_obs}, \code{cluster_of_obs}, \code{hh_of_obs}, and \code{times} added (or overwritten).
}
\description{
Creates a list object that follows the structure required by
 the exposure models that are implemented in STAN.
}
\details{
This function takes as input a `wide` data frame and extracts from it the information needed for fitting the exposure model in STAN.
}
\seealso{
\code{\link{sample_exposure_model}}, \code{\link{create_standata_outcome}}
}
