% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exposure_simulation.R
\name{expsim_update_parameter}
\alias{expsim_update_parameter}
\title{Update Exposure Model Parameters}
\usage{
expsim_update_parameter(
  obj,
  param,
  level = c("group", "cluster", "unit", "observation", "correlation", "time"),
  type = c("mean", "sd", "re", "fn"),
  value = NULL,
  draw = is.null(value)
)
}
\arguments{
\item{obj}{'expsim' object to update (see \code{\link{create_exposure_simulation_skeleton}}).}

\item{param}{character string giving name of parameter to update. If not provided, parameter name is taken from \code{level} and \code{type}.}

\item{level}{character string of model level being updated. Must be one of "group", "cluster", "unit", "observation", or "time". See Details.}

\item{type}{character string giving type of parameter being updated. See Details.}

\item{value}{value(s) to which parameter should be set}

\item{draw}{if \code{TRUE}, then the parameter value is sampled instead of being set to \code{value}}
}
\description{
Sets or overwrites data-generating parameters for an exposure model object
}
\details{
Only specific combinations of \code{level} and \code{type} are allowed. They are:
\itemize{
\item \code{"group"}: \code{"mean"} sets 'etaG'
\item \code{"cluster"}: \code{"sd"} and \code{"re"} set 'sigK' and 'reK', respectively
\item \code{"unit"}: \code{"sd"} and \code{"re"} set 'sigI' and 'reI', respectively
\item \code{"observation"}: \code{"sd"} sets 'sigW'.
\item \code{"time"}: \code{"fn"} sets 'timefn'. To update the times themselves, use \code{\link{sim_update_times}}.
}
After setting and/or sampling values via this funciton, use \code{\link{expsim_sample_observations}} to sample the exposure values.
}
\examples{
# Create simulation structure
es <- create_exposure_simulation_skeleton_parallel(ngroups=2, nunits=10)

# No standard deviation parameter set by default
es$structure$sigK
# Add standard deviation parameter value
es <- expsim_update_parameter(es, level="cluster", type="sd", value=1)
es$structure$sigK

# Sample random effect values
es <- expsim_update_parameter(es, level="cluster", type="re", draw=TRUE)

# Add time function
es <- expsim_update_parameter(es, level="time", type="fn", value=function(w) 2*cos(w))
}
\seealso{
Other exposure simulation functions: 
\code{\link{create_exposure_simulation_skeleton}()},
\code{\link{expsim_sample_observations}()},
\code{\link{sim_update_times}()}
}
\concept{exposure simulation functions}
