% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_outcome_model.R
\name{get_fitted_ERC}
\alias{get_fitted_ERC}
\alias{plot_fitted_ERC}
\alias{center_ERC}
\title{Extract fitted Exposure-Response Curve (ERC)}
\usage{
get_fitted_ERC(
  standata,
  stanfit,
  beta_post,
  exprange = c(0, 100),
  ciband = 0.95,
  inclInterceptUncertainty = TRUE,
  inclIntercept = FALSE,
  intercept_prop = c("equal", "obs"),
  ...
)

plot_fitted_ERC(
  obj,
  incS = NULL,
  expERC = TRUE,
  ylab = "Relative Risk",
  xlab = "Exposure",
  ribbon = FALSE
)

center_ERC(obj, ref_exposure = min(obj$exposure))
}
\arguments{
\item{standata}{\code{standata_outcome} object used to fit the model}

\item{stanfit}{fitted outcome model from \code{\link{sample_outcome_model}}. Posterior samples of 'beta' are extracted from this object.}

\item{beta_post}{vector or matrix of posterior samples for 'beta' parameter. Only needed if \code{stanfit} not provided.}

\item{exprange}{range of exposure values over which to compute the curve}

\item{ciband}{width of credible interval band}

\item{inclInterceptUncertainty}{Should intercept uncertainty be included uncertainty estimates? See details for more information.}

\item{inclIntercept}{Should the intercept term be included in the curve?}

\item{intercept_prop}{Proportions used in calculating the "average" intercept. Defaults to equal proportions for each study (\code{"equal"}) and can be set to be proportional to the number of observations in each study (\code{"obs"}). A vector of proportions can also be given.}

\item{...}{Passed to \code{\link{seq}} to control sequence of exposure values.}

\item{obj}{Data frame containing \code{exposure}, \code{mean}, \code{low}, and \code{high}. Typically generated from \code{\link{get_fitted_ERC}}.}

\item{incS}{If model has a single curve but with different selections of intercept uncertainty, this indicates which choice of uncertainty to use. Defaults to the last column of the curve in \code{obj}, which is typically the averaged intercept uncertainty. If multiple curves are fit, this selects which curve(s) is plotted.}

\item{expERC}{Should the fitted curve be exponentiated (TRUE) or not (FALSE).}

\item{ylab}{String providing y-axis label.}

\item{xlab}{String providing x-axis label.}

\item{ribbon}{Should the uncertainty be represented as a filled ribbon (TRUE) or lines without fill (FALSE).}

\item{ref_exposure}{Exposure concentration at which the spline should be shifted to have value 0 on log scale (value 1 on exponentiated scale). The largest exposure value less than \code{ref_exposure} is used as the new reference concentration.}
}
\description{
Computes the exposure-response curve from a fitted outcome model
}
\details{
This function creates a data frame containing the values of the exposure-response curve over a given range of values. The output is designed for easy plotting.
Currently, the fitted curve is plotted based upon the posterior means of the parameters, with the uncertainty bands based upon quantiles.

Uncertainty from the intercepts is included in the confidence bands by default, since this corresponds to a common interpretation of such intervals. This requires picking a single value for the intercept. For models fit to multiple studies, a separate set of uncertainty will be created for each study. Additionally, a set of results corresponding to "average" intercept is created. The \code{intercept_prop} argument controls the relative contribution of the intercepts from each model.
}
\seealso{
\code{\link{sample_outcome_model}}
}
