% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_outcome_model.R
\name{get_fitted_pm_curve}
\alias{get_fitted_pm_curve}
\alias{plot_fitted_pm_curve}
\title{Extract fitted PM curve}
\usage{
get_fitted_pm_curve(standata, stanfit, beta_post, pmrange = c(0, 100),
  ciband = 0.95, inclInterceptUncertainty = TRUE)

plot_fitted_pm_curve(obj, incS = NULL, ylab = "Relative Risk",
  xlab = "PM Concentration (ug/m3)")
}
\arguments{
\item{standata}{\code{standata_outcome} object used to fit the model}

\item{stanfit}{fitted outcome model from \code{\link{sample_outcome_model}}. Posterior samples of 'beta' are extracted from this object.}

\item{beta_post}{vector or matrix of posterior samples for 'beta' parameter. Only needed if \code{stanfit} not provided.}

\item{pmrange}{range of values over which to compute the curve}

\item{ciband}{width of credible interval band}

\item{inclInterceptUncertainty}{include intercept uncertainty in the output?}

\item{obj}{Data frame containing \code{pm}, \code{mean}, \code{low}, and \code{high}. Typically generated from \code{\link{get_fitted_pm_curve}}.}

\item{incS}{Which studies to include. Default of NULL plots all studies.}

\item{ylab}{String providing y-axis label.}

\item{xlab}{String providing x-axis label.}
}
\description{
Computes the concentration-response curve from a fitted outcome model
}
\seealso{
\code{\link{sample_outcome_model}}
}
