% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_outcome_simulation.R
\name{outsim_sample_observations}
\alias{outsim_sample_observations}
\title{Sample Simulated Outcome Observations}
\usage{
outsim_sample_observations(obj)
}
\arguments{
\item{obj}{the outcome simulation object for which to sample}
}
\description{
Draws a sample of outcome observations using provided model parameters
}
\examples{
skeleton <- create_outcome_simulation_skeleton_parallel()
outsim_sample_observations(skeleton)
}
\seealso{
\code{\link{expsim_sample_observations}}

Other outcome simulation functions: 
\code{\link{create_outcome_simulation_skeleton}()},
\code{\link{outsim_update_parameter}()},
\code{\link{sim_update_times}()}
}
\concept{outcome simulation functions}
