% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_outcome_simulation.R
\name{outsim_update_parameter}
\alias{outsim_update_parameter}
\alias{outsim_update_atrisk}
\alias{outsim_update_covariate}
\alias{outsim_update_set_parameter}
\alias{outsim_update_exp_splines}
\title{Update Outcome model parameters}
\usage{
outsim_update_parameter(
  obj,
  param,
  level = c("study", "unit", "time", "covariate", "exposure", "obs"),
  type = c("mean", "sd", "re", "fn", "coef"),
  value = NULL,
  draw = is.null(value)
)

outsim_update_atrisk(obj, nT = 1)

outsim_update_covariate(
  obj,
  covariate = c("x", "Z"),
  ncol = 1,
  value = NULL,
  draw = is.null(value),
  verbose = TRUE,
  mean = 0,
  sd = 1
)

outsim_update_set_parameter(obj, param = c("beta", "gamma"), value = NULL)

outsim_update_exp_splines(obj, df = 1, fn = "ns", ...)
}
\arguments{
\item{obj}{'outsim' object to update}

\item{param}{character string giving name of parameter to update. If not provided, parameter name is taken from \code{level} and \code{type}}

\item{level}{character string of model level being updated. See details.}

\item{type}{character string giving type of parameter being updated. See details.}

\item{value}{value(s) to which parameter should be set}

\item{draw}{if \code{TRUE}, then the parameter value is sampled instead of being set to \code{value}}

\item{nT}{time at risk value to set}

\item{covariate}{character string indicating which covariate to update. currently only 'x' and 'Z'.}

\item{ncol}{number of columns in covariate matrix. only used if \code{draw=TRUE}.}

\item{verbose}{Set to \code{FALSE} to suppress messages.}

\item{mean}{Mean of normal distribution from which to draw.}

\item{sd}{SD of normal dsitirbtion from which to draw.}

\item{df}{Degrees of freedom for exposure spline.}

\item{fn}{Name of spline-generating function. Defaults to \code{\link[splines]{ns}}}

\item{...}{Additional arguments passed to \code{\link{create_spline_matrix}}.}
}
\description{
Sets or overwrites outcome model parameters
}
\details{
The function \code{outsim_update_parameter} can be used to set or update the values that determine the outcome observations. Only specific combinations of \code{level} and \code{type} are allowed. They are:
\itemize{
\item \code{"study"}: \code{"mean"}  sets the overall intercept (on the scale of the linear predictor), denoted as \code{beta0}.
\item \code{"unit"}: \code{"sd"} sets the standard deviation for individual random effects; \code{"re"} samples the random effects themselves.
\item \code{"unit"}: \code{"sd"} sets the standard deviation for individual random effects; \code{"re"} samples the random effects themselves.
\item \code{"time"}: \code{"mean"} sets 'timefn'. To update the times themselves, use \code{\link{sim_update_times}}.
\item \code{"covariate"}: \code{"coef"} sets the covariate (\code{Z}) coefficients.  To update the covariate values themselves, use \code{\link{outsim_update_covariate}}.
\item \code{"exposure"}: \code{"fn"} sets 'xfn', the underlying exposure-response function (on the scale of the linear predictor). To update the exposure values themselves, use \code{\link{outsim_update_covariate}}.
}
}
\examples{
skel <- create_outcome_simulation_skeleton_parallel()
outsim_update_parameter(obj=skel, param='beta0', value=2)
}
\seealso{
\code{\link{expsim_update_parameter}}

\code{\link{create_spline_matrix}}

Other outcome simulation functions: 
\code{\link{create_outcome_simulation_skeleton}()},
\code{\link{outsim_sample_observations}()},
\code{\link{sim_update_times}()}
}
\concept{outcome simulation functions}
