% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exposure_model.R
\name{plot_exposure_means_bytime}
\alias{plot_exposure_means_bytime}
\alias{plot_exposure_means_boxplot}
\title{Plot Exposure Model Summaries}
\usage{
plot_exposure_means_bytime(
  stanfit,
  standata,
  include_time = TRUE,
  include_reI = TRUE,
  exp_transform = FALSE,
  group_names = paste0("Group ", 1:standata$G)
)

plot_exposure_means_boxplot(
  stanfit,
  standata,
  include_time = TRUE,
  include_reI = TRUE,
  exp_transform = FALSE,
  one_per_person = !include_time,
  group_names = paste0("Group ", 1:standata$G)
)
}
\arguments{
\item{stanfit}{Fitted STAN model object containing posterior samples of parameters; from \code{\link{sample_exposure_model}}.}

\item{standata}{List containing model structure information corresponding to \code{stanfit}.
Typically a list of class \code{standata_exposure}. However, it may be any list (or data frame) containing
\code{cluster_of_obs} (if \code{etaK} was sampled in \code{stanfit}), \code{group_of_obs}
(if \code{etaK} was not sampled in \code{stanfit}),  \code{unit_of_obs} (if \code{include_reI=TRUE}),
 and \code{Mt} (if \code{include_time=TRUE}). This may be useful when calculating fitted means for records
  at different times from the observations used to fit the model.}

\item{include_time}{Logical indicator of whether time should be included.}

\item{include_reI}{Logical indicator of whether the unit-level random effect should be included.}

\item{exp_transform}{Logical indicator of whether concentrations should be exponentiated.}

\item{group_names}{Names of groups to use in plot labels}

\item{one_per_person}{Logical indicator of whether a single observation per unit should be included. If \code{FALSE} (the default), then
all observations are included in the boxplot. Setting to \code{TRUE} is most useful when \code{include_time=FALSE} and there is no group crossover, in which case all fitted posterior means for a single unit are the same.}
}
\description{
Different plots of the concentrations modelled in the exposure model
}
\details{
Uses \code{\link[ggplot2]{ggplot}} to create plots of concentrations using posterior means of model parameters. The graphical object is returned and can be customized if needed.

Both functions are wrappers around \code{\link{compute_fitted_mean()}} and first call that function to compute the posterior mean for each observation.
}
\examples{
data(casedataA)
data(casedataB)
outcome_combo_data <- create_standata_outcome(datalist=list(casedataA, casedataB),
                                              xdf=4,
                                              xfnargs=list(Boundary.knots=c(5, 200)))
outcome_combo_data <- add_priors(outcome_combo_data,
sigmaI=c(0, 0.1))

outcome_combo_mod_fit <- sample_outcome_model(outcome_combo_data,
B=2000,
cores=4)

plot_exposure_means_bytime(stanfit = outcome_combo_mod_fit, standata = outcome_combo_data)
}
