% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_outcome_model.R
\name{sample_outcome_model}
\alias{sample_outcome_model}
\title{Sample Outcome Model}
\usage{
sample_outcome_model(standata, B = 2000, warmup = B, chains = 4,
  control = list(adapt_delta = 0.9, max_treedepth = 12),
  multiple_exposure_curves = FALSE, restrictBeta = FALSE, ...)
}
\arguments{
\item{standata}{An object of class `standata_outcome`, typically created from \code{\link{create_standata_outcome}}.}

\item{B}{Number of post-warmup iterations.}

\item{warmup}{Number of warmup iterations.}

\item{chains}{Number of chains to sample.}

\item{control}{List provided as the \code{control} argument of \code{\link[rstan]{sampling}}}

\item{multiple_exposure_curves}{Logical indicating whether exposure response curves should be estimated separately by study.}

\item{restrictBeta}{Logical indicating that model should be fit that forces a positive value for the exposure spline coefficients.}

\item{...}{Additional arguments passed to \code{\link[rstan]{sampling}}.}
}
\description{
Samples from the posterior distribution of outcome model parameters, using STAN
}
\seealso{
\link{create_standata_outcome}, \link{outsim_sample_observations}
}
\author{
Joshua Keller
}
