% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exposure_model.R
\name{compute_fitted_mean}
\alias{compute_fitted_mean}
\title{Compute fitted concentration}
\usage{
compute_fitted_mean(stanfit, standata, include_time = TRUE,
  include_reH = TRUE, exp_transform = FALSE, add = FALSE, ...)
}
\arguments{
\item{stanfit}{Fitted STAN model object containing posterior samples of parameters}

\item{standata}{List containing model structure information corresponding to \code{stanfit}.
Typically a list of class \code{standata_exposure}. However, it may be any list (or data frame) containing
\code{cluster_of_obs} (if \code{etaK} was sampled in \code{stanfit}), \code{group_of_obs}
(if \code{etaK} was not sampled in \code{stanfit}),  \code{hh_of_obs} (if \code{include_reH=TRUE}),
 and \code{Ht} (if \code{include_time=TRUE}). This may be useful when calculating fitted means for records
  at different times from the observations used to fit the model.}

\item{include_time}{Logical indicator of whether time should be included.}

\item{include_reH}{Logical indicator of whether the household-level random effect should be included.}

\item{exp_transform}{Logical indicator of whether concentrations should be exponentiated.}

\item{add}{Logical indicator of whether the modeled means should be added to \code{standata} or just returned directly (the default).}

\item{...}{Additional arugments passed to \code{\link[rstan]{extract}}.}
}
\description{
Calculates the posterior mean exposure concentrations.
}
\seealso{
\code{\link{plot_exposure_means_bytime}}
}
