% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exposure_simulation.R
\name{create_exposure_simulation_skeleton}
\alias{create_exposure_simulation_skeleton}
\alias{create_exposure_simulation_skeleton_parallel}
\title{Create List Structure for Simulation}
\usage{
create_exposure_simulation_skeleton(design = "parallel", ...)

create_exposure_simulation_skeleton_parallel(ngroups = 1, nclusters = 1,
  nhouseholds = 1, nobs = 1, verbose = TRUE)
}
\arguments{
\item{design}{String providing the study design. Currently only \code{'parallel'} is implemented.}

\item{...}{Additional arguments passed to the design-specific functions.}

\item{ngroups}{Number of groups.}

\item{nclusters}{Number of clusters. Either a single value or a vector of length \code{ngroups}.}

\item{nhouseholds}{Number of households in each cluster. If length 1, repeated for all clusters.}

\item{nobs}{Number of observations in each household. If length 1, repeated for all households and if length \code{sum(nclusters)}, repeated for all households within each cluster.}

\item{verbose}{Logical indicator for message printing.}
}
\value{
A list containing two sublists: \code{structure}, which contains settings and parameters for generating data, and \code{standata} which contains the study data in a format for sampling via STAN.
}
\description{
These functions construct lists containing the structure and data for simulating exposure and outcome data.
}
\details{
These functions create the blank structure of groups, clusters, households, and subject.
Once parameters are set via \code{\link{expsim_update_parameter}} and related functions, then data can be sampled via 
\code{\link{expsim_sample_observations}}
and posterior parameter estimates obtained from \code{\link{sample_exposure_model}}.
}
\seealso{
\code{\link{create_outcome_simulation_skeleton}}

Other exposure simulation functions: \code{\link{expsim_sample_observations}},
  \code{\link{expsim_update_parameter}},
  \code{\link{sim_update_time_splines}}
}
