% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exposure_simulation.R
\name{create_exposure_simulation_skeleton}
\alias{create_exposure_simulation_skeleton}
\alias{create_exposure_simulation_skeleton_parallel}
\title{Create List Structure for Simulations}
\usage{
create_exposure_simulation_skeleton(design = "parallel", ...)

create_exposure_simulation_skeleton_parallel(
  ngroups = 1,
  nclusters = 1,
  nunits = 1,
  nobs = 1,
  verbose = TRUE
)
}
\arguments{
\item{design}{String providing the study design. Currently only \code{"parallel"} is implemented.}

\item{...}{Additional arguments passed to the design-specific functions.}

\item{ngroups}{Number of groups.}

\item{nclusters}{Number of clusters. Either a single value or a vector of length \code{ngroups}.}

\item{nunits}{Number of units (e.g. households) in each cluster. If length 1, repeated for all clusters.}

\item{nobs}{Number of observations in each unit. If length 1, repeated for all units and if length \code{sum(nclusters)}, repeated for all units within each cluster.}

\item{verbose}{Logical indicator for message printing.}
}
\value{
A list containing two sublists: \code{structure}, which contains settings and parameters for generating data, and \code{standata} which contains the study data as a \code{standata_exposure} obejct for sampling via \code{\link{sample_exposure_model}}.
}
\description{
These functions construct lists containing the structure for simulating exposure data and the objects needed for fitting the model to the data.
}
\details{
These functions create the blank structure of groups, clusters, and units (e.g. households).
Once parameters are set via \code{\link{expsim_update_parameter}} and related functions, then data can be sampled via
\code{\link{expsim_sample_observations}}
and posterior parameter estimates obtained from \code{\link{sample_exposure_model}}.
See \code{\link{sample_exposure_model}} for a mathematical description of the model.
}
\examples{
# Create simulation structure
es <- create_exposure_simulation_skeleton_parallel(ngroups=2, nunits=10)
# No standard deviation parameter set by default
es$structure$sigK
# Add standard deviation parameter value
es <- expsim_update_parameter(es, level="cluster", type="sd", value=1)
es$structure$sigK
# Sample the model
es2 <- expsim_sample_observations(es)
}
\seealso{
\code{\link{create_outcome_simulation_skeleton}}, \code{\link{sample_exposure_model}}

Other exposure simulation functions: 
\code{\link{expsim_sample_observations}()},
\code{\link{expsim_update_parameter}()},
\code{\link{sim_update_time_splines}()}
}
\concept{exposure simulation functions}
