% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exposure_model.R
\name{create_standata_exposure}
\alias{create_standata_exposure}
\title{Create List for Fitting Exposure Model via STAN}
\usage{
create_standata_exposure(
  data = NULL,
  group = data$group,
  conc = data$conc,
  unit_id = data$unit_id,
  clust_id = data$clust_id,
  time = data$time,
  Mt = NULL,
  log_transform = FALSE,
  return_addition = FALSE
)
}
\arguments{
\item{data}{Optional data frame containing data in a long format.}

\item{group}{Vector of group assignments. See Details.}

\item{conc}{Vector of exposure concentrations.}

\item{unit_id}{Vector identifying the distinct unit (e.g. person, household) for each observation.}

\item{clust_id}{Vector identifying cluster membership for each observation.}

\item{time}{Vector of times corresponding to \code{conc} values.}

\item{Mt}{Matrix of splines values for time to include in the model. Defaults to NULL, and can be added later via \code{\link{add_spline_time}}.}

\item{log_transform}{Should the concentration value be log-transformed?}

\item{return_addition}{See 'Value'.}
}
\value{
A list of class \code{standata_exposure} that contains the following:
\itemize{
\item {\code{G} -- Number of groups}
\item {\code{K} -- Number of clusters.}
\item{\code{n} Number of households}
\item{\code{N} -- Number of observations}
\item{\code{cluster_of_obs} -- Integer providing cluster number of each observation}
\item{\code{group_of_obs} -- Integer providing group number of each observation}
\item{\code{hh_of_obs} -- Integer providing household number of each observation}
\item{\code{w} -- Concentration observations}
\item{\code{times} -- The date/time of each observation. This is not used directly in model fitting, but contained in the object to facilitate plotting and other summaries.}
}
If \code{return_addition=TRUE}, then a two-element list is returned. The first element
is the \code{standata_exposure} object described above. The second element is a modified version of \code{data}, with the variables \code{group_of_obs}, \code{cluster_of_obs}, \code{hh_of_obs}, and \code{times} added (or overwritten).
}
\description{
Creates a list object that follows the structure required by
 the exposure models that are implemented in STAN.
}
\details{
This function takes as input a `long` data frame and extracts from it the information needed for fitting the exposure model in STAN.

The values of \code{group}, \code{unit_id}, and \code{clust_id} are converted to factor and then coerced to integer. The ordering of groups (units, clusters) in the output object will depend on the default ordering introduced by \code{\link{factor}}.
}
\seealso{
\code{\link{sample_exposure_model}}, \code{\link{create_standata_outcome}}
}
