% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exposure_simulation.R
\name{expsim_sample_observations}
\alias{expsim_sample_observations}
\title{Sample Simulated Exposure Observations}
\usage{
expsim_sample_observations(obj)
}
\arguments{
\item{obj}{exposure simulation object, created by \code{\link{create_exposure_simulation_skeleton}}.}
}
\value{
An object of class \code{expsim}, with updated values of \code{obj$standata$w} and \code{obj$structure$meanW}.
}
\description{
Draws a sample of exposure observations using the specified model parameters
}
\details{
This function assumes that \code{etaG}, \code{sigW}, \code{reI}, and \code{timefn} have been set. Optionally, \code{sigK} should be set if a cluster random effect is to be included. These can be set with \code{\link{expsim_update_parameter}}.
}
\seealso{
Other exposure simulation functions: 
\code{\link{create_exposure_simulation_skeleton}()},
\code{\link{expsim_update_parameter}()},
\code{\link{sim_update_time_splines}()}
}
\concept{exposure simulation functions}
