% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exposure_simulation.R
\name{expsim_sample_observations}
\alias{expsim_sample_observations}
\title{Sample Simulated Exposure Observations}
\usage{
expsim_sample_observations(obj)
}
\arguments{
\item{obj}{exposure simualtion object, created by \code{\link{create_exposure_simulation_skeleton}}.}
}
\description{
Draws a sample of exposure observations using the specified model parameters
}
\details{
This function assumes that \code{etaK} (or at least \code{etaG}), \code{sigW}, \code{reH}, and \code{timefn} have been set. The latter defaults to zero, but the others must be set with \code{\link{expsim_update_parameter}}. If these have not been set, then this will set \code{w} to a vector of \code{NA}.
}
\seealso{
Other exposure simulation functions: \code{\link{create_exposure_simulation_skeleton}},
  \code{\link{expsim_update_parameter}},
  \code{\link{sim_update_time_splines}}
}
\concept{exposure simulation functions}
