% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_outcome_model.R
\name{get_fitted_ERC}
\alias{get_fitted_ERC}
\alias{plot_fitted_ERC}
\title{Extract fitted Exposure-Response Curve (ERC)}
\usage{
get_fitted_ERC(
  standata,
  stanfit,
  beta_post,
  exprange = c(0, 100),
  ciband = 0.95,
  inclInterceptUncertainty = TRUE,
  inclIntercept = FALSE,
  ...
)

plot_fitted_ERC(obj, incS = NULL, ylab = "Relative Risk", xlab = "Exposure")
}
\arguments{
\item{standata}{\code{standata_outcome} object used to fit the model}

\item{stanfit}{fitted outcome model from \code{\link{sample_outcome_model}}. Posterior samples of 'beta' are extracted from this object.}

\item{beta_post}{vector or matrix of posterior samples for 'beta' parameter. Only needed if \code{stanfit} not provided.}

\item{exprange}{range of exposure values over which to compute the curve}

\item{ciband}{width of credible interval band}

\item{inclInterceptUncertainty}{Should intercept uncertainty be included uncertainty estimates? See details for more information.}

\item{inclIntercept}{Should the intercept term be included in the curve?}

\item{...}{Passed to \code{\link{seq}} to control sequence of exposure values.}

\item{obj}{Data frame containing \code{exposure}, \code{mean}, \code{low}, and \code{high}. Typically generated from \code{\link{get_fitted_ERC}}.}

\item{incS}{Which studies to include. Default of NULL plots all studies.}

\item{ylab}{String providing y-axis label.}

\item{xlab}{String providing x-axis label.}
}
\description{
Computes the exposure-response curve from a fitted outcome model
}
\details{
This function creates a data frame containing the values of the exposure-response curve over a given range of values. The output is designed for easy plotting.
Currently, the fitted curve is plotted based upon the posterior means of the parameters, with the uncertainty bands based upon quantiles.
}
\seealso{
\code{\link{sample_outcome_model}}
}
