% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exposure_model.R
\name{plot_exposure_means_bytime}
\alias{plot_exposure_means_bytime}
\alias{plot_exposure_means_boxplot}
\title{Plot Exposure Model Summaries}
\usage{
plot_exposure_means_bytime(
  stanfit,
  standata,
  include_time = TRUE,
  include_reI = TRUE,
  exp_transform = FALSE,
  group_names = paste0("Group ", 1:standata$G)
)

plot_exposure_means_boxplot(
  stanfit,
  standata,
  include_time = TRUE,
  include_reI = TRUE,
  exp_transform = FALSE,
  one_per_person = !include_time,
  group_names = paste0("Group ", 1:standata$G)
)
}
\arguments{
\item{stanfit}{Fitted STAN model object containing posterior samples of parameters; from \code{\link{sample_exposure_model}}.}

\item{standata}{List containing model structure information corresponding to \code{stanfit}.
Typically a list of class \code{standata_exposure}. However, it may be any list (or data frame) containing
\code{cluster_of_obs} (if \code{etaK} was sampled in \code{stanfit}), \code{group_of_obs}
(if \code{etaK} was not sampled in \code{stanfit}),  \code{unit_of_obs} (if \code{include_reI=TRUE}),
and \code{Mt} (if \code{include_time=TRUE}). This may be useful when calculating fitted means for records
at different times from the observations used to fit the model.}

\item{include_time}{Logical indicator of whether time should be included.}

\item{include_reI}{Logical indicator of whether the unit-level random effect should be included.}

\item{exp_transform}{Logical indicator of whether concentrations should be exponentiated.}

\item{group_names}{Names of groups to use in plot labels}

\item{one_per_person}{Logical indicator of whether a single observation per unit should be included. If \code{FALSE} (the default), then
all observations are included in the boxplot. Setting to \code{TRUE} is most useful when \code{include_time=FALSE} and there is no group crossover, in which case all fitted posterior means for a single unit are the same.}
}
\description{
Different plots of the concentrations modelled in the exposure model
}
\details{
Uses \code{\link[ggplot2]{ggplot}} to create plots of concentrations using posterior means of model parameters. The graphical object is returned and can be customized if needed.

Both functions are wrappers around \code{\link{compute_fitted_mean}} and first call that function to compute the posterior mean for each observation.
}
\examples{
#Create simulated data
exposure_standata <- create_standata_exposure(group=rep(1:3, each=24),
conc=rnorm(n=72,mean=rep(c(0, 2, 4), each=23)),
unit_id=rep(1:18, each=4),
time=rep(1:4, times=18))

#Add natural splines with 3 df
Mt <- create_spline_matrix(x=exposure_standata$time, df=3, fn="ns")

#Add priors and spline
exposure_standata <- add_spline_time(exposure_standata, Mt=Mt)
exposure_standata <- add_priors(exposure_standata, sigmaI=c(0, 0.5))

#Sample from posterior distribution using STAN
exposure_mod_fit <- sample_exposure_model(exposure_standata, B=200, chains=2)

#Plot long-term means
plot_exposure_means_bytime(stanfit = exposure_mod_fit, standata = exposure_standata)
}
