% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exposure_model.R
\name{plot_exposure_means_bytime}
\alias{plot_exposure_means_bytime}
\alias{plot_exposure_means_boxplot}
\title{Plot Exposure Model Summaries}
\usage{
plot_exposure_means_bytime(
  stanfit,
  standata,
  include_time = TRUE,
  include_reI = TRUE,
  exp_transform = FALSE,
  group_names = paste0("Group ", 1:standata$G)
)

plot_exposure_means_boxplot(
  stanfit,
  standata,
  include_time = TRUE,
  include_reI = TRUE,
  exp_transform = FALSE,
  one_per_person = !include_time,
  group_names = paste0("Group ", 1:standata$G)
)
}
\arguments{
\item{stanfit}{Fitted STAN model object containing posterior samples of parameters}

\item{standata}{List containing model structure information corresponding to \code{stanfit}.
Typically a list of class \code{standata_exposure}. However, it may be any list (or data frame) containing
\code{cluster_of_obs} (if \code{etaK} was sampled in \code{stanfit}), \code{group_of_obs}
(if \code{etaK} was not sampled in \code{stanfit}),  \code{unit_of_obs} (if \code{include_reI=TRUE}),
 and \code{Mt} (if \code{include_time=TRUE}). This may be useful when calculating fitted means for records
  at different times from the observations used to fit the model.}

\item{include_time}{Logical indicator of whether time should be included.}

\item{include_reI}{Logical indicator of whether the unit-level random effect should be included.}

\item{exp_transform}{Logical indicator of whether concentrations should be exponentiated.}

\item{group_names}{Names of groups to use in plot labels}

\item{one_per_person}{Logical indicator of whether a single observation per person should be included. If \code{FALSE} (the default), then
all observations are included in the boxplot.}
}
\description{
Different plots of the concentrations modelled in the exposure model
}
\details{
Uses \code{\link[ggplot2]{ggplot}} to create plots of concentrations using posterior means of model parameters. The graphical object is returned and can be customized if needed.
}
