% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exposure_model.R
\name{sample_exposure_model}
\alias{sample_exposure_model}
\title{Sample Exposure Model}
\usage{
sample_exposure_model(
  standata,
  B = 1000,
  warmup = B,
  chains = 4,
  control = list(adapt_delta = 0.9, max_treedepth = 12),
  ...
)
}
\arguments{
\item{standata}{An object of class \code{standata_exposure}, typically created from \code{\link{create_standata_exposure}}. It should also include information on prior distributions, which can be added via \code{\link{add_priors}}}

\item{B}{Number of post-warmup iterations.}

\item{warmup}{Number of warmup iterations.}

\item{chains}{Number of chains to sample.}

\item{control}{List provided as the \code{control} argument of \code{\link[rstan]{sampling}}}

\item{...}{Additional arguments passed to \code{\link[rstan]{sampling}}.}
}
\description{
Samples from the posterior distribution of the exposure model, using STAN
}
\details{
This fits a hierarchical model designed for modeling exposure concentrations from a study that has irregularly-spaced, clustered measurements. The model for an observation \eqn{w_{gkit}} from exposure group \eqn{g}, cluster (e.g. neighborhood) \eqn{k}, unit (e.g. household) \eqn{i}, and time \eqn{t} is:
\deqn{w_{gkit} = \eta_g + \alpha_{0k} + \alpha_{1i}+ H(t)*\theta}
Here, \eqn{\eta_g} is a group-specific mean, \eqn{alpha_{0k}} is a cluster-specific random effect, \eqn{alpha_{1i}} is a unit-specific random effect, and \eqn{\theta} are coefficients for a temporal spline \eqn{H(t)}.
}
\examples{
# Create simulated data
exp_data <- create_standata_exposure(group=rep(1, 10),
                                     conc=rnorm(10),
                                     unit_id=rep(0:1, 5),
                                     time=runif(10))
# Add default priors
exp_data <- add_priors(exp_data,
                       sigmaI=c(0, 0.1))
# Fit the model
exp_mod_fit <- sample_exposure_model(exp_data)
print(exp_mod_fit, pars=c("muW", "reI_raw", "etaG_raw"), include=FALSE)
}
\seealso{
\code{\link{create_standata_exposure}}, \code{\link{sample_outcome_model}}, \code{\link{add_priors}}, \code{\link{compute_fitted_mean}}
}
