% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_outcome_model.R
\name{sample_outcome_model}
\alias{sample_outcome_model}
\title{Sample Outcome Model}
\usage{
sample_outcome_model(
  standata,
  B = 1000,
  warmup = B,
  chains = 4,
  control = list(adapt_delta = 0.9, max_treedepth = 12),
  multiple_exposure_curves = FALSE,
  restrictBeta = FALSE,
  continuous = TRUE,
  ...
)
}
\arguments{
\item{standata}{An object of class `standata_outcome`, typically created from \code{\link{create_standata_outcome}}.}

\item{B}{Number of post-warmup iterations.}

\item{warmup}{Number of warmup iterations.}

\item{chains}{Number of chains to sample.}

\item{control}{List provided as the \code{control} argument of \code{\link[rstan]{sampling}}}

\item{multiple_exposure_curves}{Logical indicating whether exposure response curves should be estimated separately by study.}

\item{restrictBeta}{Logical indicating that model should be fit that forces a positive value for the exposure spline coefficients.}

\item{continuous}{If the outcome is continuous (TRUE) or binary (FALSE)}

\item{...}{Additional arguments passed to \code{\link[rstan]{sampling}}.}
}
\description{
Samples from the posterior distribution of outcome model parameters, using STAN
}
\examples{
data(casedataA)
outcome_dataA <- create_standata_outcome(data=casedataA)
outcome_dataA <- add_priors(outcome_dataA,
                            sigmaI=c(0, 0.1))
outcome_mod_fit1 <- sample_outcome_model(outcome_dataA)
print(outcome_mod_fit1, pars=c("reI_raw", "reI","mui"), include=FALSE)
}
\seealso{
\link{create_standata_outcome}, \link{outsim_sample_observations}
}
\author{
Joshua Keller
}
